+++ src/main/java/twentyfive/twentyfiveadapter/generic/ecommerce/models/dinamic/stat/IngredientStatCategory.java
@@ -2,5 +2,5 @@ package twentyfive.twentyfiveadapter.generic.ecommerce.models.dinamic.stat;
 
 public class IngredientStatCategory {
     private String idIngredient;
-    private int totalTimeUsed;
+    private long totalTimeUsed;
 }
